refactor: adjust privacy code
