+++ pygwalker/__init__.py
@@ -6,7 +6,7 @@ init_logging()
 import logging
 
 from pygwalker.utils.randoms import rand_str as __rand_str
-from pygwalker_utils.config import get_config as __get_config
+from pygwalker.services.global_var import GlobalVarManager
 
 __version__ = "0.3.10a0"
 __hash__ = __rand_str()
@@ -16,7 +16,7 @@ from pygwalker.api.gwalker import GWalker
 from pygwalker.api.html import to_html
 from pygwalker.data_parsers.base import FieldSpec
 
-if __get_config('privacy')[0] == 'offline':
+if GlobalVarManager.privacy == 'offline':
     logging.getLogger(__name__).info("Running in offline mode. There might be newer releases available. Please check at https://github.com/Kanaries/pygwalker or https://pypi.org/project/pygwalker.")
 else:
     from pygwalker.services.check_update import check_update
