--- pygwalker_utils/defaults.json
@@ -1,3 +1,0 @@
-{
-    "privacy": "meta"
-}
